feat: separated video-call into home and video
feat: passed id using the router

feat: added new Home page
